# Add option to use smaller AES tables (table sizes reduced by 6144 bytes)