@sbutcher-arm 2016-04-16
Thanks for the contribution!

I think we'd like to accept this once I've given it a fuller review, but to accept your current contributions or any future contributions we will need a Contributor’s Licence Agreement (CLA) signed or authorised by yourself. You can [find an agreement to sign here](https://www.mbed.com/en/about-mbed/contributor-license-agreements/), which can be signed and returned to us, or you could create an mbed account and [accept a slightly different agreement here with a click through](https://developer.mbed.org/contributor_agreement/) if this is a personal contribution. 

Thanks for your understanding, and apologies it's taken so long to come back to you since you made the contribution.