@mpg 2018-03-27
Note: I did some time tests on a FRDM-K64F board (based on a Cortex-M4 core) with Mbed OS (using the online compiler), and here are the results for running time of `mbedtls_aes_self_test(0)` (with CBC, CFB and CTR enabled) in milliseconds, depending of whether tables are stored in ROM or RAM, and whether we use 4 tables or 1. In the RAM case, the first line is with tables no initialised yet, and the second with tables initialised.

```
           4       1
RAM     1248    1244
        1248    1244
ROM     2614    2563
```

The results surprise me:
- apparently the time it takes to generate the tables in negligible
- apparently the code runs slightly faster with fewer tables
However I checked them many times and can't find a flaw with the measurement.

If the results are correct, then I would suggest making 1 table the new default.