@Patater 2018-03-29
Hi @jkivilin

Thanks very much for your contribution. We've merged your work into the development branch via commit 8645e908255694e8d3c3f1c04b439cc66cdd35a3.

Thanks!