@hanno-arm 2018-03-29
Hi @jkivilin, 

apologies for taking such a long time to consider and merge your PR - it's without question that a delay as the one experienced here is unfavourable for multiple reasons, and we're actively working on reducing the time to review and integrate community contributions at the moment.

Thanks again for your contribution,
Kind regards,
Hanno